+++ packages/server/src/server/__tests__/index.test.ts
@@ -49,7 +49,7 @@ describe('server', () => {
         'x-faasjs-request-id': 'test',
         'x-headers': 'x-x',
         'access-control-expose-headers':
-          'content-type,authorization,x-faasjs-request-id,x-faasjs-timing-pending,x-faasjs-timing-processing,x-faasjs-timing-total',
+          'content-type, authorization, x-faasjs-request-id, x-faasjs-timing-pending, x-faasjs-timing-processing, x-faasjs-timing-total, accept-encoding',
       },
       statusCode: 200,
       body: { data: 'hello' },
@@ -124,7 +124,7 @@ describe('server', () => {
       headers: {
         'access-control-allow-credentials': 'true',
         'access-control-allow-headers':
-          'x-x, accept-encoding, x-y, content-type, authorization, x-faasjs-request-id, x-faasjs-timing-pending, x-faasjs-timing-processing, x-faasjs-timing-total',
+          'content-type, authorization, x-faasjs-request-id, x-faasjs-timing-pending, x-faasjs-timing-processing, x-faasjs-timing-total, x-x, accept-encoding, x-y',
         'access-control-allow-methods': 'OPTIONS, POST',
         'access-control-allow-origin': '*',
       },