+++ CHANGELOG.md
@@ -7,8 +7,10 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 [`v7.0.0 (2025-04-06)`](https://github.com/faasjs/faasjs/compare/v6.12.3...v7.0.0)
 
 - `@faasjs/server`
+  - [Break] All header names are lowercase.
   - [Break] 3rd of middleware argument changes to `MiddlewareContext`.
   - [Feature] Add port in use notification.
+  - [Feature] Add `beforeHandle` to `Server`.
 
 [`v6.12.3 (2025-04-06)`](https://github.com/faasjs/faasjs/compare/v6.11.1...v6.12.3)